id types changed from long to int

+++ Source/Source/Horeb.MoneySaver.Domain/Modules/FinanceModule/MonthlyBalanceEnquiry.cs
@@ -13,8 +13,8 @@ namespace Horeb.Domain.FinanceModule
     {
         public Decimal MonthlyEndingBalance { get; set; }
 
-        public long WalletId { get; set; }
+        public int WalletId { get; set; }
 
-        public long MonthlyPeriodId { get; set; }
+        public int MonthlyPeriodId { get; set; }
     }
 }
